+++ eras/alonzo/impl/src/Cardano/Ledger/Alonzo/PlutusScriptApi.hs
@@ -53,6 +53,7 @@ import Cardano.Ledger.BaseTypes (ProtVer (pvMajor), natVersion)
 import Cardano.Ledger.Binary (DecCBOR (..), EncCBOR (..))
 import Cardano.Ledger.Binary.Coders
 import Cardano.Ledger.Language (BinaryPlutus (..), Language (..), Plutus (..))
+import Cardano.Ledger.TreeDiff (ToExpr)
 import Cardano.Ledger.TxIn (TxIn (..))
 import Cardano.Ledger.UTxO (EraUTxO (..), ScriptsProvided (..), UTxO (..))
 import Cardano.Slotting.EpochInfo (EpochInfo)
@@ -98,6 +99,8 @@ data CollectError era
   | NoCostModel !Language
   | BadTranslation !(TranslationError (EraCrypto era))
   deriving (Generic)
+
+instance ToExpr (TxCert era) => ToExpr (CollectError era)
 
 deriving instance (Era era, Eq (TxCert era)) => Eq (CollectError era)
 deriving instance (Era era, Show (TxCert era)) => Show (CollectError era)