+++ eras/conway/impl/src/Cardano/Ledger/Conway/Rules/Certs.hs
@@ -59,6 +59,7 @@ import Cardano.Ledger.Shelley.Rules (
   validateDelegationRegistered,
   validateZeroRewards,
  )
+import Cardano.Ledger.TreeDiff (ToExpr)
 import Control.Monad.Trans.Reader (asks)
 import Control.State.Transition.Extended (
   Embed (..),
@@ -94,6 +95,10 @@ data ConwayCertsPredFailure era
   | -- | CERT rule subtransition Failures
     CertFailure !(PredicateFailure (EraRule "CERT" era))
   deriving (Generic)
+
+instance
+  ToExpr (PredicateFailure (EraRule "CERT" era)) =>
+  ToExpr (ConwayCertsPredFailure era)
 
 deriving stock instance
   Eq (PredicateFailure (EraRule "CERT" era)) =>