+++ eras/conway/impl/src/Cardano/Ledger/Conway/Rules/Deleg.hs
@@ -38,6 +38,7 @@ import Cardano.Ledger.Core (Era (EraCrypto), EraPParams, EraRule, PParams)
 import Cardano.Ledger.Credential (Credential)
 import Cardano.Ledger.Keys (KeyRole (Staking))
 import Cardano.Ledger.Shelley.LedgerState (DState (..))
+import Cardano.Ledger.TreeDiff (ToExpr)
 import qualified Cardano.Ledger.UMap as UM
 import Control.DeepSeq (NFData)
 import Control.Monad (forM_)
@@ -69,6 +70,8 @@ data ConwayDelegPredFailure era
   | DRepAlreadyRegisteredForStakeKeyDELEG !(Credential 'Staking (EraCrypto era))
   | WrongCertificateTypeDELEG
   deriving (Show, Eq, Generic)
+
+instance ToExpr (ConwayDelegPredFailure era)
 
 instance NoThunks (ConwayDelegPredFailure era)